Create Bib superclass
